Removed graph_context
Graph context was removed. All code from it was implemented into graph.py. Tests updated accordingly
