+++ tests/unit/commands/check/test_graph.py
@@ -1,22 +1,253 @@
 from unittest import TestCase
-from unittest.mock import Mock, patch
+from unittest.mock import Mock, MagicMock, patch
 
-from samcli.commands.check.resources.graph import CheckGraph
+from samcli.commands.check.resources.graph import CheckGraph, _check_input
 
 
 class TestGraph(TestCase):
-    def test_class(self):
-        entry_point_mock = Mock()
+    @patch("samcli.commands.check.resources.graph.CheckGraph")
+    def test_generate(self, patch_graph):
+        graph_mock = Mock()
+        patch_graph.return_value = graph_mock
+
+        graph = CheckGraph([])
+
+        graph._make_connections = Mock()
+        graph._find_entry_points = Mock()
+
+        graph.generate()
+
+        graph._make_connections.assert_called_once()
+        graph._find_entry_points.assert_called_once()
+
+    def test_handle_lambda_permissions(self):
+        self_mock = Mock()
+
+        permission_object = {}
+        function_name_mock = Mock()
+        function_object_mock = Mock()
+        function_object_mock.parents = []
+        source_name_mock = Mock()
+        source_object_mock = Mock()
+        source_object_mock.children = []
+
+        permission_object = {
+            "Properties": {
+                "FunctionName": {"Ref": function_name_mock},
+                "SourceArn": {"Fn::Sub": [None, {"__ApiId__": {"Ref": source_name_mock}}]},
+            }
+        }
+
+        self_mock._lambda_functions = {function_name_mock: function_object_mock}
+        self_mock._api_gateways = {source_name_mock: source_object_mock}
+
+        permission_mock = Mock()
+        permission_mock.resource_object = permission_object
+
+        self_mock._lambda_permissions.values.return_value = [permission_mock]
+
+        CheckGraph._handle_lambda_permissions(self_mock)
+
+        self.assertEqual(function_object_mock.parents[0], source_object_mock)
+        self.assertEqual(source_object_mock.children[0], function_object_mock)
+
+        self_mock._lambda_functions = {}
+
+        with self.assertRaises(Exception):
+            CheckGraph._handle_lambda_permissions(self_mock)
+
+    @patch("samcli.commands.check.resources.graph.DynamoDB")
+    def test_handle_event_source_mappings(self, patch_dynamo):
+        self_mock = Mock()
+        event_mock = Mock()
+        function_name_mock = Mock()
+        function_object_mock = Mock()
+        function_object_mock = Mock()
+        function_object_mock.parents = []
+        source_object_mock = Mock()
+        source_object_mock.children = []
+
+        patch_dynamo.return_value = source_object_mock
+
+        source_name = "arn:aws:dynamodb:region:id:path:num"
+
+        event_object = {"Properties": {"FunctionName": {"Ref": function_name_mock}, "EventSourceArn": source_name}}
+
+        event_mock.resource_object = event_object
+
+        self_mock._event_source_mappings.values.return_value = [event_mock]
+        self_mock._lambda_functions = {function_name_mock: function_object_mock}
+
+        self_mock._dynamo_db_tables = {}
+
+        CheckGraph._handle_event_source_mappings(self_mock)
+
+        self.assertEqual(function_object_mock.parents[0], source_object_mock)
+        self.assertEqual(source_object_mock.children[0], function_object_mock)
+
+    def test_find_entry_points(self):
+        self_mock = Mock()
+        self_mock.entry_points = []
+
+        function_object_mock = Mock()
+        api_object_mock = Mock()
+        dynamo_object_mock = Mock()
+
+        function_object_mock.parents = []
+        api_object_mock.parents = []
+        dynamo_object_mock.parents = []
+
+        self_mock._lambda_functions.values.return_value = [function_object_mock]
+        self_mock._api_gateways.values.return_value = []
+        self_mock._dynamo_db_tables.values.return_value = []
+
+        CheckGraph._find_entry_points(self_mock)
+
+        self.assertEqual(self_mock.entry_points[0], function_object_mock)
+
+        self_mock._api_gateways.values.return_value = [api_object_mock]
+        CheckGraph._find_entry_points(self_mock)
+
+        self.assertEqual(self_mock.entry_points[-1], api_object_mock)
+
+        self_mock._dynamo_db_tables.values.return_value = [dynamo_object_mock]
+        CheckGraph._find_entry_points(self_mock)
+
+        self.assertEqual(self_mock.entry_points[-1], dynamo_object_mock)
+
+    def test_make_connections(self):
+        self_mock = Mock()
+
+        self_mock._handle_lambda_permissions = Mock()
+        self_mock._handle_event_source_mappings = Mock()
+        self_mock._handle_iam_roles = Mock()
+
+        CheckGraph._make_connections(self_mock)
+
+        self_mock._handle_lambda_permissions.assert_called_once()
+        self_mock._handle_event_source_mappings.assert_called_once()
+        self_mock._handle_iam_roles.assert_called_once()
+
+    def test_handle_iam_roles(self):
+        lambda_function_name = ""
+        lambda_function_mock = Mock()
+        policy_mock = Mock()
+
+        policies = [policy_mock]
+
+        properties = {"ManagedPolicyArns": policies}
+
+        self_mock = Mock()
+
+        self_mock._lambda_functions = {lambda_function_name: lambda_function_mock}
+        self_mock._get_properties.return_value = properties
+        self_mock._dynamo_policies = policies
+        self_mock._make_connection_from_policy = Mock()
+
+        CheckGraph._handle_iam_roles(self_mock)
+
+        self_mock._get_properties.assert_called_once_with(lambda_function_mock)
+        self_mock._make_connection_from_policy.assert_called_once_with("AWS::DynamoDB::Table", lambda_function_name)
+
+    def test_get_properties(self):
+        self_mock = Mock()
+
+        properties_mock = Mock()
+
+        iam_role_object = {"Properties": properties_mock}
+
+        role_mock = Mock()
+        role_mock.resource_object = iam_role_object
+
+        lambda_function_role_name = Mock()
+        lambda_function_resource_object = {"Properties": {"Role": {"Fn::GetAtt": [lambda_function_role_name]}}}
+
+        self_mock._iam_roles = {lambda_function_role_name: role_mock}
+
+        lambda_function_mock = Mock()
+        lambda_function_mock.resource_object = lambda_function_resource_object
+
+        result = CheckGraph._get_properties(self_mock, lambda_function_mock)
+
+        # Test return value
+        self.assertEqual(result, properties_mock)
+
+        # test error raised
+        self_mock._iam_roles = {}
+
+        with self.assertRaises(Exception):
+            CheckGraph._get_properties(self_mock, lambda_function_mock)
+
+    def test_make_connection_from_policy(self):
         resource_mock = Mock()
+        resource_mock.parents = []
+        resources_selected = [resource_mock]
 
-        lambda_functions = []
+        child_resource_type = "AWS::DynamoDB::Table"
+        lambda_function_name = Mock()
+        lambda_function_mock = Mock()
+        lambda_function_mock.children = []
 
-        key = "key"
+        self_mock = Mock()
+        self_mock._ask_dynamo_connection_question.return_value = resources_selected
+        self_mock._lambda_functions = {lambda_function_name: lambda_function_mock}
 
-        graph = CheckGraph(lambda_functions)
+        CheckGraph._make_connection_from_policy(self_mock, child_resource_type, lambda_function_name)
 
-        graph.entry_points.append(entry_point_mock)
-        graph.resources_to_analyze[key] = resource_mock
+        self.assertEqual(lambda_function_mock.children[0], resource_mock)
+        self.assertEqual(resource_mock.parents[0], lambda_function_mock)
 
-        self.assertEqual(entry_point_mock, graph.entry_points[0])
-        self.assertEqual(resource_mock, graph.resources_to_analyze[key])
+    @patch("samcli.commands.check.resources.graph._check_input")
+    @patch("samcli.commands.check.resources.graph.click")
+    def test_ask_dynamo_connection_question(self, patch_click, patch_check):
+        lambda_function_name = ""
+        dynamo_table_name = ""
+
+        user_input_mock = Mock()
+        selected_resource_mock = Mock()
+
+        self_mock = Mock()
+        self_mock._dynamo_db_tables = {dynamo_table_name: selected_resource_mock}
+
+        choice = 1
+        user_choices = [choice]
+
+        patch_click.prompt.return_value = user_input_mock
+
+        patch_check.return_value = True, user_choices
+
+        result = CheckGraph._ask_dynamo_connection_question(self_mock, lambda_function_name)
+
+        patch_check.assert_called_once_with(user_input_mock, 1)
+        patch_click.prompt.assert_called_once()
+
+        self.assertEqual(result, [selected_resource_mock])
+
+    @patch("samcli.commands.check.resources.graph.click")
+    def test_check_input(self, patch_click):
+        # All valid input from user
+        user_input = "1"
+        max_item_number = 1
+
+        bool_result, choices_result = _check_input(user_input, max_item_number)
+
+        self.assertEqual(bool_result, True)
+        self.assertEqual(choices_result, [1])
+
+        # Non-int entered
+        user_input = "One"
+        max_item_number = 1
+
+        bool_result, choices_result = _check_input(user_input, max_item_number)
+
+        self.assertEqual(bool_result, False)
+        self.assertEqual(choices_result, [])
+
+        # Outside of range
+        user_input = "1:2:3:4"
+        max_item_number = 3
+
+        bool_result, choices_result = _check_input(user_input, max_item_number)
+
+        self.assertEqual(bool_result, False)
+        self.assertEqual(choices_result, [])
